# openapi-functions.yaml
...
swagger: "2.0"
info:
  title: alerts description
  description: API Gateway with a Google Cloud Functions backend
  version: 1.0.0
schemes:
  - https
produces:
  - application/json
paths:
...
  /bot/{botId}:
    delete:
      summary: delete bot
      operationId: deleteBot
      x-google-backend:
        address: https://us-central1-cald-ads-qa.cloudfunctions.net
        path_translation: APPEND_PATH_TO_ADDRESS
      parameters:
        - in: path
          name: botId
          required: true
          type: integer
      responses:
        "200":
          description: A successful response
          schema:
            type: string
    patch:
      summary: patch bot
      operationId: patchBot
      x-google-backend:
        address: https://us-central1-cald-ads-qa.cloudfunctions.net
        path_translation: APPEND_PATH_TO_ADDRESS
      parameters:
        - in: path
          name: botId
          required: true
          type: integer
      responses:
        "200":
          description: A successful response
          schema:
            type: string
  /bot/run/{botId}:
    post:
      summary: run bot
      operationId: runBot
      x-google-backend:
        address: https://us-central1-cald-ads-qa.cloudfunctions.net
        path_translation: APPEND_PATH_TO_ADDRESS
      parameters:
        - in: path
          name: botId
          required: true
          type: integer
      responses:
        "200":
          description: A successful response
          schema:
            type: string
...